Fix account loading twice because of incorrent IsRefreshingUsage
IsRefreshing should not be bound to true because this triggers a refresh. Previously it was bound to IsLoading, thus when this was set to true, a refresh was triggered. So when the page was initializing, it was refreshing twice simultaneously. The first time the wallet opens the CPU is hit hard to decrypt the wallet, so doing it twice was killing performance.

Now, the loading happens once per initialization and refreshing works as expected. Also, the default loading spinner is visible once again.

Last but not least, the view initialisation doesn't invoke the viewmodel's Initialize method during OnAppearing. Instead, IsRefreshing is bound to true after 10 milliseconds thus triggering the 1st load after the UI appears.