Use a factory to resolve IBolService instances
This is using a scope factory to re-use the same scope until a wallet is closed. When this happens, the factory disposes the scope so future requests to resolve IBolService happen under a new scope. This way it is ensured the underlying VM is always recreated to properly support opening a new wallet.