Send a message about wallet creation so BolService can be disposed before re-use
This will ensure that instances needed after wallet creation will properly get access to the new wallet data. This is needed because an IBolService instance is already created to be used to validate the new codename during wallet creation, thus it cannot be re-used after wallet creation, since it uses an IOptions<BolWallet> instance that doesn't contain the new wallet data, so the service needs to be re-initialized.