Register more SDK services as transient
RpcClient depends on an IOptions<BolConfig> instance and is registered as scoped by default. This causes issues after closing a wallet and opening another, since subsequent usage will use the same BolConfig so transactions are failing or are unpredictable.

TransactionService is also registered as scoped and depends on IRpcClient, so making both of them transient overcomes the issue.